requirements.txt
pyfiglet
mysql_connector_python
tqdm
wget
